server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
python-dotenv==1.0.1
httpx==0.28.1
itsdangerous==2.2.0
jinja2==3.1.5
python-multipart==0.0.20
